Fixed performance counter test, fix for issue #319
