--- cv32/tests/programs/custom/counters/test.yaml
@@ -1,5 +1,0 @@
-name: counters
-uvm_test: uvmt_cv32_firmware_test_c
-description: >
-    Simple counters sanity test
-
