+++ cv32/tests/programs/custom/hpmcounter_hazard_test/test.yaml
@@ -1,0 +1,5 @@
+name: hpmcounter_hazard_test
+uvm_test: uvmt_cv32_firmware_test_c
+description: >
+    HPM Counter Hazard Test (donated by Arjan Bink)
+
